Оптимизировал интерфейс приложения
