--- src/components/place-with-sorting/index.ts
@@ -1,3 +1,0 @@
-import PlaceWithSorting from './place-with-sorting';
-
-export default PlaceWithSorting;
