+++ src/components/private-route/private-route.tsx
@@ -1,12 +1,15 @@
 import { Navigate } from 'react-router-dom';
 import { AppRoute, AuthStatus } from '../../settings';
+import { useAppSelector } from '../../hooks';
+import { getAuthStatus } from '../../store/user-process/selectors';
 
 type PrivateRouteProps = {
   children: JSX.Element;
-  authStatus: AuthStatus;
 };
 
-export default function PrivateRoute({ children, authStatus }: PrivateRouteProps): JSX.Element {
+export default function PrivateRoute({ children }: PrivateRouteProps): JSX.Element {
+  const authStatus = useAppSelector(getAuthStatus);
+
   return authStatus === AuthStatus.Auth ? (
     children
   ) : (
